requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
gspread>=5.12.0
google-auth>=2.23.0
google-api-python-client>=2.100.0
